Feature/new implementations

+++ src/main/java/com/example/advancedtaskmanagement/task/task_progress/TaskProgressRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TaskProgressRepository extends JpaRepository<TaskProgress, Integer> {
+public interface TaskProgressRepository extends JpaRepository<TaskProgress, Long> {
 
     Optional<TaskProgress> findByIdAndIsDeletedFalse(Long id);
 }
